Rename player_components_factory.h to factory.cc
